requirements.txt
Django==5.0.6
pytest-django==4.8.0
celery[redis]==5.4.0
psycopg[binary]==3.1.19
djangorestframework==3.15.1
markdown==3.5.2
django-filter==24.2
django-celery-results==2.5.1
requests==2.31.0
cryptography==42.0.7
